crew-runtime/requirements.txt
fastapi==0.96.0
uvicorn[standard]==0.23.1
grpcio==1.56.0
grpcio-tools==1.56.0
openai==1.76.0
pytest==7.4.0
